Add button boundary feature
When a button was touched, then the users finger moved/dragged outside the
button area, and then subsequently released their finger... the button
would fire regardless of whether the pointer was inside or outside the
buttons bounds.

To allow the user to escape an incorectly touched button this feature
extends the buttons perimeter/bounds by a default value of 30 pixels.

This mechanism allows the user to move/drag their finger outside the
extended boundary/perimeter, then release their finger, causing the
button to not fire. Only when the pointer is within the extended
boundary at the time the pointer is released will the button fire.

Resolves: #000007