Add alternative way to close dialog
To close the popup dialog the user had to previously
tap the same button used to open it. Often users
would attempt to tap the main content area to close
the dialog - yet nothing would happen.

To improve the UX the popup dialog is now closed
when they tap/click on the main content area.

Resolves: 000009